Struct rusoto_polly::SynthesizeSpeechInput

pub struct SynthesizeSpeechInput {
    pub lexicon_names: Option<Vec<String>>,
    pub output_format: String,
    pub sample_rate: Option<String>,
    pub speech_mark_types: Option<Vec<String>>,
    pub text: String,
    pub text_type: Option<String>,
    pub voice_id: String,
}

Fields
lexicon_names: Option<Vec<String>>

List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice. For information about storing lexicons, see PutLexicon.

output_format: String

The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

sample_rate: Option<String>

The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", and "22050". The default value is "22050".

Valid values for pcm are "8000" and "16000" The default value is "16000".

speech_mark_types: Option<Vec<String>>

The type of speech marks returned for the input text.

text: String

Input text to synthesize. If you specify ssml as the TextType, follow the SSML format for the input text.

text_type: Option<String>

Specifies whether the input text is plain text or SSML. The default value is plain text. For more information, see Using SSML.

voice_id: String

Voice ID to use for the synthesis. You can get a list of available voice IDs by calling the DescribeVoices operation.