Trait rusoto_polly::Polly

pub trait Polly {
    fn delete_lexicon(
        &self, 
        input: DeleteLexiconInput
    ) -> RusotoFuture<DeleteLexiconOutput, DeleteLexiconError>;

    fn describe_voices(
        &self, 
        input: DescribeVoicesInput
    ) -> RusotoFuture<DescribeVoicesOutput, DescribeVoicesError>;

    fn get_lexicon(
        &self, 
        input: GetLexiconInput
    ) -> RusotoFuture<GetLexiconOutput, GetLexiconError>;

    fn get_speech_synthesis_task(
        &self, 
        input: GetSpeechSynthesisTaskInput
    ) -> RusotoFuture<GetSpeechSynthesisTaskOutput, GetSpeechSynthesisTaskError>;

    fn list_lexicons(
        &self, 
        input: ListLexiconsInput
    ) -> RusotoFuture<ListLexiconsOutput, ListLexiconsError>;

    fn list_speech_synthesis_tasks(
        &self, 
        input: ListSpeechSynthesisTasksInput
    ) -> RusotoFuture<ListSpeechSynthesisTasksOutput, ListSpeechSynthesisTasksError>;

    fn put_lexicon(
        &self, 
        input: PutLexiconInput
    ) -> RusotoFuture<PutLexiconOutput, PutLexiconError>;

    fn start_speech_synthesis_task(
        &self, 
        input: StartSpeechSynthesisTaskInput
    ) -> RusotoFuture<StartSpeechSynthesisTaskOutput, StartSpeechSynthesisTaskError>;

    fn synthesize_speech(
        &self, 
        input: SynthesizeSpeechInput
    ) -> RusotoFuture<SynthesizeSpeechOutput, SynthesizeSpeechError>;
}

Trait representing the capabilities of the Amazon Polly API. Amazon Polly clients implement this trait.

Required methods

fn delete_lexicon(
&self,
input: DeleteLexiconInput
) -> RusotoFuture<DeleteLexiconOutput, DeleteLexiconError>

Deletes the specified pronunciation lexicon stored in an AWS Region. A lexicon which has been deleted is not available for speech synthesis, nor is it possible to retrieve it using either the GetLexicon or ListLexicon APIs.

For more information, see Managing Lexicons.

fn describe_voices(
&self,
input: DescribeVoicesInput
) -> RusotoFuture<DescribeVoicesOutput, DescribeVoicesError>

Returns the list of voices that are available for use when requesting speech synthesis. Each voice speaks a specified language, is either male or female, and is identified by an ID, which is the ASCII version of the voice name.

When synthesizing speech ( SynthesizeSpeech ), you provide the voice ID for the voice you want from the list of voices returned by DescribeVoices.

For example, you want your news reader application to read news in a specific language, but giving a user the option to choose the voice. Using the DescribeVoices operation you can provide the user with a list of available voices to select from.

You can optionally specify a language code to filter the available voices. For example, if you specify en-US, the operation returns a list of all available US English voices.

This operation requires permissions to perform the polly:DescribeVoices action.

fn get_lexicon(
&self,
input: GetLexiconInput
) -> RusotoFuture<GetLexiconOutput, GetLexiconError>

Returns the content of the specified pronunciation lexicon stored in an AWS Region. For more information, see Managing Lexicons.

fn get_speech_synthesis_task(
&self,
input: GetSpeechSynthesisTaskInput
) -> RusotoFuture<GetSpeechSynthesisTaskOutput, GetSpeechSynthesisTaskError>

Retrieves a specific SpeechSynthesisTask object based on its TaskID. This object contains information about the given speech synthesis task, including the status of the task, and a link to the S3 bucket containing the output of the task.

fn list_lexicons(
&self,
input: ListLexiconsInput
) -> RusotoFuture<ListLexiconsOutput, ListLexiconsError>

Returns a list of pronunciation lexicons stored in an AWS Region. For more information, see Managing Lexicons.

fn list_speech_synthesis_tasks(
&self,
input: ListSpeechSynthesisTasksInput
) -> RusotoFuture<ListSpeechSynthesisTasksOutput, ListSpeechSynthesisTasksError>

Returns a list of SpeechSynthesisTask objects ordered by their creation date. This operation can filter the tasks by their status, for example, allowing users to list only tasks that are completed.

fn put_lexicon(
&self,
input: PutLexiconInput
) -> RusotoFuture<PutLexiconOutput, PutLexiconError>

Stores a pronunciation lexicon in an AWS Region. If a lexicon with the same name already exists in the region, it is overwritten by the new lexicon. Lexicon operations have eventual consistency, therefore, it might take some time before the lexicon is available to the SynthesizeSpeech operation.

For more information, see Managing Lexicons.

fn start_speech_synthesis_task(
&self,
input: StartSpeechSynthesisTaskInput
) -> RusotoFuture<StartSpeechSynthesisTaskOutput, StartSpeechSynthesisTaskError>

Allows the creation of an asynchronous synthesis task, by starting a new SpeechSynthesisTask. This operation requires all the standard information needed for speech synthesis, plus the name of an Amazon S3 bucket for the service to store the output of the synthesis task and two optional parameters (OutputS3KeyPrefix and SnsTopicArn). Once the synthesis task is created, this operation will return a SpeechSynthesisTask object, which will include an identifier of this task as well as the current status.

fn synthesize_speech(
&self,
input: SynthesizeSpeechInput
) -> RusotoFuture<SynthesizeSpeechOutput, SynthesizeSpeechError>

Synthesizes UTF-8 input, plain text or SSML, to a stream of bytes. SSML input must be valid, well-formed SSML. Some alphabets might not be available with all the voices (for example, Cyrillic might not be read at all by English voices) unless phoneme mapping is used. For more information, see How it Works.

Loading content...