Crate rusoto_polly

Amazon Polly is a web service that makes it easy to synthesize speech from text.

The Amazon Polly service provides API operations for synthesizing high-quality speech from plain text and Speech Synthesis Markup Language (SSML), along with managing pronunciations lexicons that enable you to get the best results for your application domain.

If you're using the service, you're probably looking for PollyClient and Polly.

Structs

DeleteLexiconInput

DeleteLexiconOutput

DescribeVoicesInput

DescribeVoicesOutput

GetLexiconInput

GetLexiconOutput

Lexicon
Provides lexicon name and lexicon content in string format. For more information, see Pronunciation Lexicon Specification (PLS) Version 1.0.

LexiconAttributes
Contains metadata describing the lexicon such as the number of lexemes, language code, and so on. For more information, see Managing Lexicons.

LexiconDescription
Describes the content of the lexicon.

ListLexiconsInput

ListLexiconsOutput

PollyClient
A client for the Amazon Polly API.

PutLexiconInput

PutLexiconOutput

SynthesizeSpeechInput

SynthesizeSpeechOutput

Voice
Description of the voice.

Enums

DeleteLexiconError
Errors returned by DeleteLexicon

DescribeVoicesError
Errors returned by DescribeVoices

GetLexiconError
Errors returned by GetLexicon

ListLexiconsError
Errors returned by ListLexicons

PutLexiconError
Errors returned by PutLexicon

SynthesizeSpeechError
Errors returned by SynthesizeSpeech

Traits

Polly
Trait representing the capabilities of the Amazon Polly API. Amazon Polly clients implement this trait.