Struct rusoto_polly::PollyClient

pub struct PollyClient { /* fields omitted */ }

A client for the Amazon Polly API.

Methods

impl PollyClient

pub fn new(region: Region) -> PollyClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> PollyClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,